fix: layout 설정 수정 및 clientLayout 생성
- 헤더/푸터가 보이지 않는 clientLayout 생성
- 헤더 잘못된 링크 경로 수정
- 404페이지 잘못된 이미지 링크 수정
