--- src/app/(user)/page.jsx
@@ -1,5 +1,0 @@
-"use client";
-
-export default function Page() {
-  return <>page</>;
-}
